Input colors
